\name{ecoinfoscrg-package}
\alias{ecoinfoscrg}
\title{ecoinfoscrg: Meta-Analysis model for the Florida Statewide Coastal Restoration Guide (SCRG)}
\description{
Includes functions and workflows for updating and predicting from the SCRG meta-analytic regression model for Living Shoreline suitability.

To learn more and view examples of package usage, please see the vignettes \code{browseVignettes(package = "dplyr")}.

For additional code and materials related to the model and SCRG project, visit \link{https://github.com/ecoinformatic/SCRG}.
}
\author{
\strong{Maintainer}: UCF Ecoinformatics Lab \email{Christen.Fleming@ucf.edu}, \email{George.Zaragosa@ucf.edu}, \email{Erika.Lin@ucf.edu}

\strong{Contributors}: Christen H. Fleming, George Zaragosa, Thomas A. Hellinger, Erika Y. Lin, Bruno Pinheiro, Nozomu Hirama}
\seealso{
\link{https://github.com/ecoinformatic/SCRG}
}

