\name{MEAN}
\alias{MEAN}
\docType{data}
\title{Mean used for standardizing predictor data}
\description{
Column means for predictor data of original studies used to build the meta-analysis model, used for standardization.
}
\format{
  The format is:
 Named num [1:21] 180.44 7.81 2.94 3.47 1.98 ...
 - attr(*, "names")= chr [1:21] "angle" "IT_Width" "Hab_W1" "Hab_W2" ...
}

\keyword{datasets}
\keyword{internal}
